+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+### 1.4.1 (2022-06-03)
+* UI Select component added.
+* Outside click helper added.
+
+
 ### 1.4.0 (2022-05-29)
 * Folders feature in progress.